requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.4
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django>=4.2.7,<4.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-extensions==3.2.3
django-silk==5.1.0
django-timezone-field==6.1.0
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt==5.3.1
gprof2dot==2022.7.29
idna==3.6
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2>=2.9.9,<3.0
pycodestyle==2.11.1
pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
drf-spectacular>=0.27.1,<0.28
uwsgi>=2.0.19,<2.0.25
